Re-wrote physics engine
Finished the work begun before initial commit: Making the whole physics engine more realistic

When two shiftable bodies collide, they distribute their momenta to come to a common speed. If one of them is not shiftable, the shiftable one changes its speed to match the unshiftable one.

Being pressed towards another body causes to get some of the other's momentum, until it is reached. Currently, this is a one-way feature, momentum is created from nowhere.

When you're pressed onto another block, you can use it to walk on. Currently, the other block is accelerated in the other direction, nor is it possible to use this feature to climb on walls when being jumped onto them and so being pressed on them.

A jump'n'run game is not playable when the player can't stop and change direciton in air. So I added a workaround which allows you to walk even if there is no body to walk on. However, the walk-feature of bodies is used to walk on moving bodies, such like chicken.