Changed the camera scrolling offset to a quarter of width/height
Previously, the space the player could move in without the camera to
scroll was determinded using a minumum border and a minumum space. This
caused the camera to wiggle when the window was too small.

Now, the free-movable space is simply a quarter of the available space,
so this even works fine on very small windows.