Created an excel file that tests random number generators for level height generation
The file uses a random number generator to get the acceleration, which is used for the speed of the height grow. A break should prevent the height from exceeding the range of [-1; 1], but this does not work. Either the number grows or shrinks too long, or you get something like a sinus graph.

Result: Random number generators as second derivation do not work.