Yogularm is playable on very small screens
Previously, the world would be resized when being drawed on smaller screens, but a meter would not be less than 50 pixels. Because of this, a player with a 800x600 screen would only see 16x12 instead of the optimal 38x18 blocks. This is a regocnizable disadvantage.

Now, this limit has been removed. Additionally, the optimal count of visible blocks has changed to 25x15 (approx. a resolution of 16:9). On a 1024x758 screen, a meter takes 50 pixels.