Reduced the physical size of the player so that you can't stand "on your hands"
Previously, the physical bounds of the player included its hands. The problem was, that when you got very close to a cliff, you would not fall down when the hands are over the floating block. Now, the hands are excluded and you'll fall down if only they are above the block.

Note: Experienced players may have to take a while for getting used to the change.