Updated physics to make platforms well-playable
Reverted collision handling to momentum-based (was, for a short time, force-based), so on a collision, the speed changes instantly. This avoids the "hopping" effect when the player is on a upwards-moving platform.

This has one exception: When the body that is being toched already moves away from the other one, the speed is not reduced. Otherwise, it would not be possible to jump on an upwards-moving platform, because the platform collides with the player and reduces its speed to platform speed.

Ground adhesion force is now calculated using the totalForce property which caches the collected forces from last frame. This is neccessary because the momenta can change during the collision checks.

Introduced the shift speed which indicates with which speed the body is being pushed. It is used for calculating the resulting jump speed. Standing on an updward-moving platform results in a higher jump speed than standing on downward-moving platforms.

The actualSpeed property incicates the distance travelled in the last frame, divided by elapsed time. It represents the visible speed and not the collected momentum that changes often.