Re-arranged texture tiles: Bugfix for some video cards
A texture has to have a width of 2^n and a height of 2^m where n and m are integers. Otherwise, the game won't work on some graphic cards. Now, all textures follow this scheme. To solve the problem with ghosty lines, borders are added inside every tile (128x128 pixels in current resolution). They are 4 pixels wide, which adds to a space of 8 pixels between two tile contents. To be flexible on resolution changes, the border is defined to be 1/32 of a tile size.