Separated the real texture objects from the world
de.yogularm.drawing.Texture now is only a container for the texture's name. The RenderContext implementation has a method bindTexture(Texture) which should load the given texture, if it hasn't been loaded yet. For this, a world can be rendered with any render context. The Game class now does not have a RenderContext field, but gets the RenderContext passed to the render() and setResolution() methods.

In android, the game is stored in a static class so that it can be restored if the app is minimized, but not terminated.