New component container: persistent and managed
The new component collection interface allows to get components in a specific range. This is used by the currently only implementation. This implementation divides the world into equally-sized sectors. If a component is added, it is either put into an existing sector or a new sector is created. The sectors are stored in a map[point > list], so large gaps don't need any memory.

These changes allow to build in two dimensions, which might be used in future. Anyway, no components are deleted, so jumping back to the beginning is no problem. Currently, about 100 blocks are iterated in the update and render loops; they are collected out of two to three sectors.